+++ RELEASENOTES.md
@@ -1,6 +1,11 @@
 # (upcoming)
 * tbd
 
+# Version 1.3
+* use iText instead of PdfSharp
+* use APGL 3 License for iText
+* allow drag and drop of image 
+
 # Version 1.2
 * Localization using .resx files
 * Check for app updates